0.1.2.Nightly.5
- Added public FastInsertByte to Writer.
- Added RpcLinks.
- Added LastPacketTick to TimeManager; does not function yet.
- Made a few write methods more performant.
- Added rounding options when using TimeManager.TimeToTicks.
- Several minor optimizations to many hotpath methods.
- Begun reworking NetworkTransform to be tick based.
- Put Mono.Cecil in it's own assembly.
- Organized some codegen helpers; did not touch any logic.
- Renamed FirstInitialize to InitializeOnce. This may throw an error during compilation if using a custom authenticator.
- TimeManager no longer sends step updates to clients which have not authenticated yet.
- NetworkTransform is now tick driven; still plans for more improvement.
- Fluidity MTUs are no longer forced and channels can no longer be overriden.
- Fixed SyncTypes sending 1 tick late.
- Made UpdateFrequency tooltip on DistanceCondition more clear.
- Fixed scene object not found error when visibility changes more than once.
- TimeManager now throws an error when multiple are spawned with AutomaticPhysics disabled.

+++ Assets/FishNet/Runtime/Object/NetworkBehaviour.SyncTypes.cs
@@ -283,10 +283,13 @@ namespace FishNet.Object
                             using (PooledWriter headerWriter = WriterPool.GetWriter())
                             {
                                 PacketId packetId = (isSyncObject) ? PacketId.SyncObject : PacketId.SyncVar;
-                                headerWriter.WriteByte((byte)packetId);
+                                headerWriter.WriteUInt16((ushort)packetId);
 
                                 PooledWriter dataWriter = WriterPool.GetWriter();
-                                dataWriter.WriteNetworkBehaviour(this);
+                                dataWriter.WriteNetworkBehaviour(this); 
+                                /* //clean try and clean this up similar to rpc changes
+                                // with unreliable lengths and what not. */
+
                                 dataWriter.WriteBytesAndSize(channelWriter.GetBuffer(), 0, channelWriter.Length);
 
                                 //If unreliable write packet length.
